\name{allcols}
\Rdversion{1.1}
\alias{allcols}
\title{allcols}
\description{## functions for calculation of loci exclusion}
\usage{
allcols(d, v = NULL, count = ncol(d))
}
\arguments{
  \item{d}{
}
  \item{v}{
}
  \item{count}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}