\name{anc.est.nicholson}
\Rdversion{1.1}
\alias{anc.est.nicholson}
\title{anc est nicholson}
\description{Ancestral allele frequency estimate given by Nicholson model.}
\usage{
anc.est.nicholson(sim, model)
}
\arguments{
  \item{sim}{
}
  \item{model}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}