\name{corr.panel}
\Rdversion{1.1}
\alias{corr.panel}
\title{corr.panel}
\description{}}
\usage{
corr.panel(x, y, ...)
}
\arguments{
  \item{x}{
}
  \item{y}{
}
  \item{\dots}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}