\name{cutoff.plot}
\Rdversion{1.1}
\alias{cutoff.plot}
\title{cutoff plot}
\description{Plot prediction counts against cutoff values.}
\usage{
cutoff.plot(cl)
}
\arguments{
  \item{cl}{Data frame from classify.loci.}
}
\details{
}
\value{The lattice plot.}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}