\name{deduce.param.label}
\Rdversion{1.1}
\alias{deduce.param.label}
\title{ Deduce parameters for plot subtitles}
\description{Summarize parameter values used in simulations.}
\usage{
deduce.param.label(lt, imp = display.params)
}
\arguments{
  \item{lt}{Data frame of results of simulation.}
  \item{imp}{Vector of column names to be searched and reported.}
}
\details{
}
\value{Text string, or NULL, to use as sub= argument for a high-level lattice function.}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}