\name{dens.several.s}
\Rdversion{1.1}
\alias{dens.several.s}
\title{dens several s}
\description{Plot density curves for ppp values for each selection type, to visualize the extent to which PPP-values can be used to indicate selection state in a simulation.}
\usage{
dens.several.s(df, ...)
}
\arguments{
  \item{df}{Data frame with columns ppp type s to be plotted.}
  \item{\dots}{
}
}
\details{
}
\value{The lattice plot.}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}