\name{display.params}
\Rdversion{1.1}
\alias{display.params}
\docType{data}
\title{display params}
\description{Default simulation parameters summarized in plot subtitles.}
\usage{display.params}
\format{
  The format is:
 chr [1:7] "populations" "popsize" "generations" "n.locus" ...
}
\details{
}
\source{
}
\references{
}
\examples{}