\name{estc.df}
\Rdversion{1.1}
\alias{estc.df}
\title{estc df}
\description{Convert list of model fit lists from nppp.over.time to a data frame suitable for plotting c values.}
\usage{
estc.df(fit.list)
}
\arguments{
  \item{fit.list}{List of model fit lists from nppp.over.time.}
}
\details{
}
\value{Data frame for plotting the simulated and theoretical values of c.}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}