\name{estc.over.time}
\Rdversion{1.1}
\alias{estc.over.time}
\title{estc over time}
\description{Plot evolution of C estimates over time. This is the differentiation parameter and is expected to increase linearly over time.}
\usage{
estc.over.time(data, ...)
}
\arguments{
  \item{data}{Data frame to plot, result of estc.df.}
  \item{\dots}{
}
}
\details{
}
\value{The ggplot2 plot.}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}