\name{fixation.endpoints}
\Rdversion{1.1}
\alias{fixation.endpoints}
\title{fixation.endpoints}
\description{CEX.LEG <- 1}
\usage{
fixation.endpoints(lf, main = "Loci are not always fixed for the same allele in each population", par.settings = list(), ancestral.color = "black", hilite.locus = NULL, ...)
}
\arguments{
  \item{lf}{
}
  \item{main}{
}
  \item{par.settings}{
}
  \item{ancestral.color}{
}
  \item{hilite.locus}{
}
  \item{\dots}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}