\name{hilite.best}
\Rdversion{1.1}
\alias{hilite.best}
\title{hilite best}
\description{panel.groups function for highlighting the best points by drawing grey lines and labeling the actual values.}
\usage{
hilite.best(x, y)
}
\arguments{
  \item{x}{
}
  \item{y}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}