\name{latticeplot}
\Rdversion{1.1}
\alias{latticeplot}
\title{latticeplot}
\description{}}
\usage{
latticeplot(f, d, FUN, alpha = 0.2, omit = NULL, groups = NULL, ...)
}
\arguments{
  \item{f}{
}
  \item{d}{
}
  \item{FUN}{
}
  \item{alpha}{
}
  \item{omit}{
}
  \item{groups}{
}
  \item{\dots}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}