\name{ltable}
\Rdversion{1.1}
\alias{ltable}
\title{ltable}
\description{##debug(deduce.param.label)}
\usage{
ltable(N, lt, fixed)
}
\arguments{
  \item{N}{
}
  \item{lt}{
}
  \item{fixed}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}