\name{make.lattice.table}
\Rdversion{1.1}
\alias{make.lattice.table}
\title{make.lattice.table}
\description{##debug(ltable)}
\usage{
make.lattice.table(N)
}
\arguments{
  \item{N}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}