\name{mypng}
\Rdversion{1.1}
\alias{mypng}
\title{mypng}
\description{## actually creates a PNG for presentations and PDFs for printout}
\usage{
mypng(title, P, subdir = "", paper = "a4r", show = "pdf")
}
\arguments{
  \item{title}{
}
  \item{P}{
}
  \item{subdir}{
}
  \item{paper}{
}
  \item{show}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}