\name{nicholsonppp-package}
\Rdversion{1.1}
\alias{nicholsonppp-package}
\docType{package}
\title{Evolution simulation and classification}
\description{After measuring gene frequencies of modern populations that have a common ancestor, we estimate ancestral allele frequency and selection state for each gene. A bayesian model is used and verified using the included simulator.}
\details{
\tabular{ll}{Package: \tab nicholsonppp\cr
Maintainer: \tab Toby Dylan Hocking <toby.hocking@etu.upmc.fr>\cr
Author: \tab Toby Dylan Hocking\cr
Version: \tab 1.0\cr
License: \tab GPL-3\cr
Title: \tab Evolution simulation and classification\cr
Description: \tab After measuring gene frequencies of modern populations that have a common ancestor, we estimate ancestral allele frequency and selection state for each gene. A bayesian model is used and verified using the included simulator.\cr
Depends: \tab lattice, animation, grid, plyr, reshape, ggplot2, latticedl\cr}
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\references{
}
\keyword{ package }
\seealso{
}
\examples{}