\name{nicholsonppp.list}
\Rdversion{1.1}
\alias{nicholsonppp.list}
\title{nicholsonppp list}
\description{Fit a model for each simulation.}
\usage{
nicholsonppp.list(sims)
}
\arguments{
  \item{sims}{List of simulations from nicholsonppp.}
}
\details{
}
\value{List of nicholson model fits.}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}