\name{pi.xyplot}
\Rdversion{1.1}
\alias{pi.xyplot}
\title{pi.xyplot}
\description{##debug(latticeplot)}
\usage{
pi.xyplot(f, d, panel = corr.panel, aspect = 1, cols = NULL, ...)
}
\arguments{
  \item{f}{
}
  \item{d}{
}
  \item{panel}{
}
  \item{aspect}{
}
  \item{cols}{
}
  \item{\dots}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}