\name{pop.colors.default}
\Rdversion{1.1}
\alias{pop.colors.default}
\docType{data}
\title{pop.colors.default}
\description{Default colors used for representing population types.}
\usage{pop.colors.default}
\format{
  The format is:
 chr [1:3] "blue" "turquoise" "red"
}
\details{
}
\source{
}
\references{
}
\examples{}