\name{ppp.df}
\Rdversion{1.1}
\alias{ppp.df}
\title{ppp df}
\description{Convert simulation and model fit lists to a data frame that can be used for plotting diagnostics for the ppp-value classifier.}
\usage{
ppp.df(sims, models)
}
\arguments{
  \item{sims}{List of simulations.}
  \item{models}{List of nicholson model fits.}
}
\details{
}
\value{Data frame suitable for plotting densities or sensitivity.}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}