\name{read.alpha.long}
\Rdversion{1.1}
\alias{read.alpha.long}
\title{read.alpha.long}
\description{read.results <- function(dvec)lapply(dvec,read.res.and.subsets)}
\usage{
read.alpha.long(f)
}
\arguments{
  \item{f}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}