\name{read.nicholson.outfile}
\Rdversion{1.1}
\alias{read.nicholson.outfile}
\title{read.nicholson.outfile}
\description{##debug(read.alpha.long)}
\usage{
read.nicholson.outfile(f)
}
\arguments{
  \item{f}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}