\name{read.res.and.subsets}
\Rdversion{1.1}
\alias{read.res.and.subsets}
\title{read.res.and.subsets}
\description{##        )}
\usage{
read.res.and.subsets(d)
}
\arguments{
  \item{d}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}