\name{read.res.tables}
\Rdversion{1.1}
\alias{read.res.tables}
\title{read.res.tables}
\description{## read all data into master data index}
\usage{
read.res.tables(d, fnames = c(alpha.est = "summary_alpha.out", alpha.sim = "RES.F.fin.txt", pi.sim = "F.init.txt", pi.est = "summary_pi.out", s = "S.LOC.txt", c = "res_c.out", ppp = "PPPval.out"))
}
\arguments{
  \item{d}{
}
  \item{fnames}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}