\name{read.results}
\Rdversion{1.1}
\alias{read.results}
\title{read.results}
\description{##debug(read.res.and.subsets)}
\usage{
read.results(dvec)
}
\arguments{
  \item{dvec}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}