\name{roc.loci}
\Rdversion{1.1}
\alias{roc.loci}
\title{roc loci}
\description{Plot ROC curves of loci classification for several s values.}
\usage{
roc.loci(cl, f = sensitivity ~ 1 - specificity | desc, ...)
}
\arguments{
  \item{cl}{Result of classify.loci.}
  \item{f}{Plot formula for xyplot.}
  \item{\dots}{
}
}
\details{
}
\value{The lattice plot.}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}