\name{run.nicholson}
\Rdversion{1.1}
\alias{run.nicholson}
\title{run.nicholson}
\description{  write.table(round(d),f,col.names=F,row.names=F,quote=F)}
\usage{
run.nicholson(d, N, f, trt.fun.name = NULL)
}
\arguments{
  \item{d}{
}
  \item{N}{
}
  \item{f}{
}
  \item{trt.fun.name}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}