\name{selection.colors.default}
\Rdversion{1.1}
\alias{selection.colors.default}
\docType{data}
\title{selection colors default}
\description{Default colors used for selection types (balancing, neutral, positive).}
\usage{selection.colors.default}
\format{
  The format is:
 chr [1:3] "#0080ff" "#ff00ff" "green"
}
\details{
}
\source{
}
\references{
}
\examples{}