\name{selection.symbols.default}
\Rdversion{1.1}
\alias{selection.symbols.default}
\docType{data}
\title{selection.symbols.default}
\description{Default symbols used for selection types (balancing, neutral, positive).}
\usage{selection.symbols.default}
\format{
  The format is:
 chr [1:3] "B" "N" "P"
}
\details{
}
\source{
}
\references{
}
\examples{}