\name{sim}
\Rdversion{1.1}
\alias{sim}
\title{sim}
\description{}
\usage{
sim(Npop = 8, Taille.POP = 1000, Ngeneration = 25, N.s.rep = 50, migrate = F, beta1 = 0.7, beta2 = 0.7, p.neutral = 1/3, s, scale = 1, shape = 1)
}
\arguments{
  \item{Npop}{
}
  \item{Taille.POP}{
}
  \item{Ngeneration}{
}
  \item{N.s.rep}{
}
  \item{migrate}{
}
  \item{beta1}{
}
  \item{beta2}{
}
  \item{p.neutral}{
}
  \item{s}{
}
  \item{scale}{
}
  \item{shape}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}