\name{sim.several.s}
\Rdversion{1.1}
\alias{sim.several.s}
\title{sim several s}
\description{Do several simulations each with only 1 s value.}
\usage{
sim.several.s(s, ...)
}
\arguments{
  \item{s}{Vector of s values.}
  \item{\dots}{
}
}
\details{
}
\value{List of selection simulations.}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}