\name{update.freq.mig}
\Rdversion{1.1}
\alias{update.freq.mig}
\title{update.freq.mig}
\description{## account with the calcul?}
\usage{
update.freq.mig(Freq.all1, F.pop, Taille.POP)
}
\arguments{
  \item{Freq.all1}{
}
  \item{F.pop}{
}
  \item{Taille.POP}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}