\name{update.freq.sel}
\Rdversion{1.1}
\alias{update.freq.sel}
\title{update.freq.sel}
\description{}
\usage{
update.freq.sel(Freq.all1, w.coef, type.pop)
}
\arguments{
  \item{Freq.all1}{
}
  \item{w.coef}{
}
  \item{type.pop}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}