\name{write.nicholson}
\Rdversion{1.1}
\alias{write.nicholson}
\title{write.nicholson}
\description{}
\usage{
write.nicholson(d, f)
}
\arguments{
  \item{d}{
}
  \item{f}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}