\name{classify.loci}
\Rdversion{1.1}
\alias{classify.loci}
\title{classify loci}
\description{Classify loci into selection or not.}
\usage{
classify.loci(res.df, xmax = 0.6, xlim = c(-0.15, xmax), ymax = 21, ylim = c(-2, ymax), cutoff = seq(0, xmax, l = 200))
}
\arguments{
  \item{res.df}{Result of ppp.df.}
  \item{xmax}{
}
  \item{xlim}{
}
  \item{ymax}{
}
  \item{ylim}{
}
  \item{cutoff}{Vector of cutoff values to use for the classifier.}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}