\name{cutoff.plot}
\Rdversion{1.1}
\alias{cutoff.plot}
\title{cutoff plot}
\description{Plot prediction counts against cutoff values.}
\usage{
cutoff.plot(cl, ylim = c(-5, 21), xlim = c(-0.15, 0.6), dens = NULL, f = percent ~ cutoff | s, ...)
}
\arguments{
  \item{cl}{Data frame from classify.loci.}
  \item{ylim}{Limits for y axis.}
  \item{xlim}{Limits for x axis.}
  \item{dens}{Optional density data to plot.}
  \item{f}{Plot formula for xyplot.}
  \item{\dots}{
}
}
\details{
}
\value{The lattice plot.}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}