\name{deduce.param.label}
\Rdversion{1.1}
\alias{deduce.param.label}
\title{deduce.param.label}
\description{display.params=c("populations",'loci','popsize','generations','p.neutral','fixed')}
\usage{
deduce.param.label(lt, imp)
}
\arguments{
  \item{lt}{
}
  \item{imp}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}