\name{display.params}
\Rdversion{1.1}
\alias{display.params}
\docType{data}
\title{display.params}
\description{##debug(fixation.endpoints)}
\usage{display.params}
\format{
  The format is:
 chr [1:6] "populations" "loci" "popsize" "generations" ...
}
\details{
}
\source{
}
\references{
}
\examples{}