\name{sim.summary.plot}
\Rdversion{1.1}
\alias{sim.summary.plot}
\title{ Simulation summary plot}
\description{Draw 3 simulation summary plots on the same screen (loci.over.time, anc.est.plot, fixation.endpoints).}
\usage{
sim.summary.plot(fr, g = 1, hilite.locus = NULL)
}
\arguments{
  \item{fr}{Data frame from a simulation.}
  \item{g}{Generation to plot in anc.est.plot and fixation.endpoints, and generation to emphasize in loci.over.time.}
  \item{hilite.locus}{Locus to highlight in the plots.}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}